[wip] update train on server.

+++ crowd_nav/train_on_server.py
@@ -11,8 +11,8 @@ import git
 import re
 from tensorboardX import SummaryWriter
 from crowd_sim.envs.utils.robot import Robot
-from crowd_nav.utils.trainer import VNRLTrainer, MPRLTrainer, TSRLTrainer, TD3RLTrainer
-from crowd_nav.utils.memory import ReplayMemory
+from crowd_nav.utils.trainer import VNRLTrainer, MPRLTrainer, TSRLTrainer, TD3RLTrainer, RGCNRLTrainer
+from crowd_nav.utils.memory import ReplayMemory, GraphReplayMemory
 from crowd_nav.utils.explorer import Explorer
 from crowd_nav.policy.policy_factory import policy_factory
 from crowd_nav.policy.reward_estimate import Reward_Estimator
@@ -20,6 +20,8 @@ from crowd_nav.policy.reward_estimate import Reward_Estimator
 import matplotlib.pyplot as plt
 from matplotlib.pyplot import plot, savefig
 import numpy as np
+episode_phase1 = 5000
+episode_phase2 = 5000
 
 def set_random_seeds(seed):
     """
@@ -33,15 +35,16 @@ def set_random_seeds(seed):
 def main(args):
     set_random_seeds(args.randomseed)
     # configure paths
-    make_new_dir = True
-    if os.path.exists(args.output_dir):
-        if args.overwrite:
-            shutil.rmtree(args.output_dir)
-        else:
-            shutil.rmtree(args.output_dir)
-    if make_new_dir:
-        os.makedirs(args.output_dir)
-        shutil.copy(args.config, os.path.join(args.output_dir, 'config.py'))
+    if args.resume is False:
+        make_new_dir = True
+        if os.path.exists(args.output_dir):
+            if args.overwrite:
+                shutil.rmtree(args.output_dir)
+            else:
+                shutil.rmtree(args.output_dir)
+        if make_new_dir:
+            os.makedirs(args.output_dir)
+            shutil.copy(args.config, os.path.join(args.output_dir, 'config.py'))
 
         # # insert the arguments from command line to the config file
         # with open(os.path.join(args.output_dir, 'config.py'), 'r') as fo:
@@ -62,7 +65,7 @@ def main(args):
     log_file = os.path.join(args.output_dir, 'output.log')
     in_weight_file = os.path.join(args.output_dir, 'in_model.pth')
     il_weight_file = os.path.join(args.output_dir, 'il_model.pth')
-    rl_weight_file = os.path.join(args.output_dir, 'rl_model.pth')
+    rl_weight_file = os.path.join(args.output_dir, 'best_val.pth')
 
     spec = importlib.util.spec_from_file_location('config', args.config)
     if spec is None:
@@ -94,6 +97,7 @@ def main(args):
     policy = policy_factory[policy_config.name]()
     if not policy.trainable:
         parser.error('Policy has to be trainable')
+    policy_config.gat.human_num = args.human_num
     policy.configure(policy_config, device)
     policy.set_device(device)
 
@@ -134,7 +138,10 @@ def main(args):
     checkpoint_interval = train_config.train.checkpoint_interval
 
     # configure trainer and explorer
-    memory = ReplayMemory(capacity)
+    if policy_config.name == "rgcn_rl":
+        memory = GraphReplayMemory(capacity)
+    else:
+        memory = ReplayMemory(capacity)
     model = policy.get_model()
     batch_size = train_config.trainer.batch_size
     optimizer = train_config.trainer.optimizer
@@ -165,18 +172,25 @@ def main(args):
                               freeze_state_predictor=train_config.train.freeze_state_predictor,
                               detach_state_predictor=train_config.train.detach_state_predictor,
                               share_graph_model=policy_config.model_predictive_rl.share_graph_model)
+    elif policy_config.name == 'rgcn_rl':
+        policy.set_action(action_dim, max_action, min_action)
+        trainer = RGCNRLTrainer(policy.actor, policy.critic, policy.state_predictor, memory, device, policy, writer,
+                              batch_size, optimizer, env.human_num, reduce_sp_update_frequency=train_config.train.reduce_sp_update_frequency,
+                              freeze_state_predictor=train_config.train.freeze_state_predictor,
+                              detach_state_predictor=train_config.train.detach_state_predictor,
+                              share_graph_model=policy_config.model_predictive_rl.share_graph_model)
     else:
         trainer = VNRLTrainer(model, memory, device, policy, batch_size, optimizer, writer)
     explorer = Explorer(env, robot, device, writer, memory, policy.gamma, target_policy=policy)
     policy.save_model(in_weight_file)
     # imitation learning
-    # if args.resume:
-    #     if not os.path.exists(rl_weight_file):
-    #         logging.error('RL weights does not exist')
-    #     policy.load_state_dict(torch.load(rl_weight_file))
-    #     model = policy.get_model()
-    #     rl_weight_file = os.path.join(args.output_dir, 'resumed_rl_model.pth')
-    #     logging.info('Load reinforcement learning trained weights. Resume training')
+    if args.resume:
+        if not os.path.exists(rl_weight_file):
+            logging.error('RL weights does not exist')
+        policy.load_state_dict(torch.load(rl_weight_file))
+        model = policy.get_model()
+        rl_weight_file = os.path.join(args.output_dir, 'resumed_rl_model.pth')
+        logging.info('Load reinforcement learning trained weights. Resume training')
     # elif os.path.exists(il_weight_file):
     #     policy.load_state_dict(torch.load(rl_weight_file))
     #     model = policy.get_model()
@@ -213,10 +227,11 @@ def main(args):
     # fill the memory pool with some RL experience
     if args.resume:
         robot.policy.set_epsilon(epsilon_end)
-        explorer.run_k_episodes(100, 'train', update_memory=True, episode=0)
+        explorer.run_k_episodes(3000, 'train', update_memory=True, episode=0)
         logging.info('Experience set size: %d/%d', len(memory), memory.capacity)
     episode = 0
     best_val_reward = -1
+    best_val_return = -1
     best_val_model = None
     # evaluate the model after imitation learning
 
@@ -241,8 +256,11 @@ def main(args):
     discom_time_in_last_interval = 0
     total_time_in_last_interval = 0
     eps_count = 0
-    # fw = open(sys_args.output_dir + '/data.txt', 'w')
-    # print("%f %f %f %f %f" % (0,0,0,0,0), file=fw)
+    fw = open(sys_args.output_dir + '/data.txt', 'w')
+    print("%f %f %f %f %f" % (0,0,0,0,0), file=fw)
+    # robot.policy.set_epsilon(epsilon_start)
+    # _, _, nav_time, sum_reward, ave_return, discom_time, total_time = \
+    #     explorer.run_k_episodes(300, 'train', update_memory=True, episode=episode)
     while episode < train_episodes:
         if args.resume:
             epsilon = epsilon_end
@@ -252,7 +270,16 @@ def main(args):
             else:
                 epsilon = epsilon_end
         robot.policy.set_epsilon(epsilon)
-
+        if episode == 0:
+            env.set_phase(0)
+        elif episode == 6000:
+            env.set_phase(1)
+        elif episode == 9000:
+            env.set_phase(2)
+        elif episode == 12000:
+            env.set_phase(3)
+        # elif episode == 12000:
+            # env.set_phase(4)
         # sample k episodes into memory and optimize over the generated memory
         _, _, nav_time, sum_reward, ave_return, discom_time, total_time = \
             explorer.run_k_episodes(sample_episodes, 'train', update_memory=True, episode=episode)
@@ -262,17 +289,17 @@ def main(args):
         nav_time__in_last_interval = nav_time__in_last_interval + nav_time
         discom_time_in_last_interval = discom_time_in_last_interval + discom_time
         total_time_in_last_interval = total_time_in_last_interval + total_time
-        interval = 100
+        interval = 50
         if eps_count % interval == 0:
-            reward_rec.append(reward_in_last_interval/100.0)
-            return_rec.append(return_in_last_interval/100.0)
-            discom_tim_rec.append(discom_time_in_last_interval/100.0)
-            nav_time_rec.append(nav_time__in_last_interval/100.0)
+            reward_rec.append(reward_in_last_interval/interval)
+            return_rec.append(return_in_last_interval/interval)
+            discom_tim_rec.append(discom_time_in_last_interval/interval)
+            nav_time_rec.append(nav_time__in_last_interval/interval)
             total_time_rec.append(total_time_in_last_interval/100.0)
-            logging.info('Train in episode %d reward in last 100 episodes %f %f %f %f %f', eps_count, reward_rec[-1],
+            logging.info('Train in episode %d reward in last %f episodes %f %f %f %f %f', interval, eps_count, reward_rec[-1],
                          return_rec[-1], discom_tim_rec[-1], nav_time_rec[-1], total_time_rec[-1])
-            # print("%f %f %f %f %f" % (reward_rec[-1], return_rec[-1], discom_tim_rec[-1], nav_time_rec[-1],
-            #                          total_time_rec[-1]), file=fw)
+            print("%f %f %f %f %f" % (reward_rec[-1], return_rec[-1], discom_tim_rec[-1], nav_time_rec[-1],
+                                     total_time_rec[-1]), file=fw)
             reward_in_last_interval = 0
             return_in_last_interval = 0
             nav_time__in_last_interval = 0
@@ -293,11 +320,11 @@ def main(args):
             trainer.update_target_model(model)
         # evaluate the model
         if episode % evaluation_interval == 0:
-            _, _, _, reward, _, _, _ = explorer.run_k_episodes(env.case_size['val'], 'val', episode=episode)
+            _, _, _, reward, average_return, _, _ = explorer.run_k_episodes(env.case_size['val'], 'val', episode=episode)
             explorer.log('val', episode // evaluation_interval)
 
-            if episode % checkpoint_interval == 0 and reward > best_val_reward:
-                best_val_reward = reward
+            if episode % checkpoint_interval == 0 and average_return > best_val_return and episode > episode_phase1 + episode_phase2:
+                best_val_return = average_return
                 best_val_model = copy.deepcopy(policy.get_state_dict())
         # test after every evaluation to check how the generalization performance evolves
             if args.test_after_every_eval:
@@ -310,19 +337,21 @@ def main(args):
             policy.save_model(save_every_checkpoint_rl_weight_file)
 
     # # test with the best val model
-    # fw.close()
+    fw.close()
     if best_val_model is not None:
         policy.load_state_dict(best_val_model)
         torch.save(best_val_model, os.path.join(args.output_dir, 'best_val.pth'))
-        logging.info('Save the best val model with the reward: {}'.format(best_val_reward))
+        logging.info('Save the best val model with the return: {}'.format(best_val_return))
     explorer.run_k_episodes(env.case_size['test'], 'test', episode=episode, print_failure=True)
 
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser('Parse configuration file')
-    parser.add_argument('--policy', type=str, default='tree_search_rl')
-    parser.add_argument('--config', type=str, default='configs/icra_benchmark/ts_separate.py')
-    parser.add_argument('--output_dir', type=str, default='data/output1')
+    parser.add_argument('--policy', type=str, default='rgcn_rl')
+    parser.add_argument('--config', type=str, default='configs/icra_benchmark/rgcnrl.py')
+    #parser.add_argument('--policy', type=str, default='td3rl')
+    #parser.add_argument('--config', type=str, default='configs/icra_benchmark/td3.py')
+    parser.add_argument('--output_dir', type=str, default='data/output')
     parser.add_argument('--overwrite', default=False, action='store_true')
     parser.add_argument('--weights', type=str)
     parser.add_argument('--resume', default=False, action='store_true')
@@ -330,11 +359,11 @@ if __name__ == '__main__':
     parser.add_argument('--debug', default=False, action='store_true')
     parser.add_argument('--test_after_every_eval', default=False, action='store_true')
     parser.add_argument('--randomseed', type=int, default=7)
-    parser.add_argument('--human_num', type=int, default=5)
+    parser.add_argument('--human_num', type=int, default=1)
     parser.add_argument('--safe_weight', type=float, default=1.0)
     parser.add_argument('--goal_weight', type=float, default=0.2)
-    parser.add_argument('--re_collision', type=float, default=-0.5)
-    parser.add_argument('--re_arrival', type=float, default=1.0)
+    parser.add_argument('--re_collision', type=float, default=-0.25)
+    parser.add_argument('--re_arrival', type=float, default=0.25)
 
 
     # arguments for GCN
